refactor(tests): split controller tests into focused modules
