+++ src/egui_app/controller/tests/common.rs
@@ -1,0 +1,43 @@
+use super::super::selection_edits::SelectionEditRequest;
+use super::super::test_support::{dummy_controller, sample_entry, write_test_wav};
+use super::super::*;
+use crate::egui_app::controller::collection_export;
+use crate::egui_app::controller::hotkeys;
+use crate::egui_app::state::{
+    DestructiveSelectionEdit, DragPayload, DragSource, DragTarget, FocusContext,
+    SampleBrowserActionPrompt, TriageFlagColumn, TriageFlagFilter, WaveformView,
+};
+use crate::sample_sources::Collection;
+use crate::sample_sources::collections::CollectionMember;
+use crate::waveform::DecodedWaveform;
+use egui;
+use hound::WavReader;
+use rand::SeedableRng;
+use rand::rngs::StdRng;
+use rand::seq::IteratorRandom;
+use std::io::Cursor;
+use std::mem;
+use std::path::{Path, PathBuf};
+use std::thread;
+use std::time::{Duration, Instant};
+use tempfile::tempdir;
+
+pub(super) fn max_sample_amplitude(path: &Path) -> f32 {
+    WavReader::open(path)
+        .unwrap()
+        .samples::<f32>()
+        .map(|s| s.unwrap().abs())
+        .fold(0.0, f32::max)
+}
+
+pub(super) fn prepare_browser_sample(
+    controller: &mut EguiController,
+    source: &SampleSource,
+    name: &str,
+) {
+    controller.sources.push(source.clone());
+    write_test_wav(&source.root.join(name), &[0.0, 0.1, -0.1]);
+    controller.wav_entries = vec![sample_entry(name, SampleTag::Neutral)];
+    controller.rebuild_wav_lookup();
+    controller.rebuild_browser_lists();
+}
